Completely refactored how keycloak is integrated into this project
