+++ package.json
@@ -62,7 +62,7 @@
   "dependencies": {
     "@craigmiller160/ajax-api": "2.1.0",
     "@craigmiller160/react-hook-form-material-ui": "1.3.0",
-    "@craigmiller160/react-keycloak": "^1.0.0-beta.1",
+    "@craigmiller160/react-keycloak": "^1.0.0-beta",
     "@craigmiller160/ts-functions": "1.6.0",
     "@date-io/date-fns": "^2.14.0",
     "@emotion/react": "^11.9.3",
